requirements.txt
click==6.6
ordereddict==1.1
python-dotenv==0.5.1
requests==2.10.0
six==1.10.0
slackclient==1.0.0
websocket-client==0.37.0
